Deleted useless files
