--- project.ui
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <author>Prathamesh Netake</author>
- <class>Form</class>
- <widget class="QWidget" name="Form">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>1088</width>
-    <height>728</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <property name="styleSheet">
-   <string notr="true">QPushButton:: hover{
-	background-color: #ffff;
-}
-
-</string>
-  </property>
- </widget>
- <resources/>
- <connections/>
-</ui>
